perf(protocol): borrow the model request instead of deep-copying it per attempt (#921)
`Engine::run_model_call` drove retries through an `FnMut` closure. Because
that closure may be called again it cannot move its inputs, so every
attempt paid `messages.to_vec()` — a deep copy of the whole conversation,
every message `String`, every tool call's `serde_json::Value`, every tool
result payload — plus `tools_schema.clone()`, every `ToolSchema` including
its full JSON parameter document. On the critical path of every model
call, including the overwhelmingly common case of exactly one attempt.

`CompletionRequest` owned both fields, so this was not fixable inside the
driver: wrapping them in `Arc` at the call site only moves the copy
earlier. The port type had to change.

## What changed

`CompletionRequestRef<'a>` — a borrowed view holding
`&[CompletionMessage]` / `&[ToolSchema]` plus the scalar params — becomes
the currency of the `Provider` port. Adapters implement `complete_ref` /
`complete_observed_ref` and serialize straight off the caller's slices.
`complete` / `complete_observed` remain as provided shims for the many
callers that genuinely own their request (reflection, ingest, judge), so
those call sites are untouched.

The view is `Copy` — slices and scalars only — which is what makes the
single-attempt path free: the retry closure holds one and hands out a
fresh copy per attempt without allocating.

Chosen over `Cow` fields on `CompletionRequest` itself: a lifetime on the
owned type infects everything that *stores* one (`AccountedCall.request`,
stella-serve's `ServerFrame`), forces `'de: 'a` on its derived
`Deserialize`, and rewrites ~110 struct-literal sites, for no additional
benefit. Every adapter already read `&req.messages` / `&req.tools` and
never moved out of either, so a view type drops in with signature-only
changes.

`CompletionRequest::as_borrowed` destructures exhaustively without `..`,
so a field added to the owned type and not threaded through the view is a
compile error rather than a parameter that silently stops reaching the
wire. The view is deliberately NOT `Serialize`: a second derive would be a
second authority on the wire shape, and the two could diverge on a
`skip_serializing_if` — exactly the drift the byte-stability contract
forbids.

## The same defect on the auxiliary path

`run_accounted_call` cloned its request per attempt for the identical
`FnMut` reason. It now borrows too — which matters most for the overflow
summarizer, whose input is a rendered span of the transcript.

## The one copy that stays

`stella-serve`'s `RemoteProvider` materializes an owned request with
`CompletionRequestRef::into_owned`: its frame crosses from the session
thread to the server runtime and outlives the borrow. Taken explicitly,
once, at the boundary that needs it — not silently once per retry attempt
inside the engine.

## Measured

`bench/request-bench`, 150k-token transcript, 60 MCP tool schemas, 200
steps, one attempt per step:

    arm            per step (bytes)   per step (allocs)   per turn (MiB)
    owned-clone             1109630                6196            211.6
    borrowed                      0                   0              0.0

## Proved

`driver/tests/zero_copy_request.rs` asserts on slice *addresses*, not
allocation counts (an allocator probe would measure the whole test
binary): a retried attempt is handed the identical `messages`/`tools`
memory as the first, and that memory is the caller's own transcript.

This also makes the prompt-cache stability contract (L-E8 / #372)
structural rather than merely observed — there is no intermediate buffer
in which the system prefix could drift, because there is no intermediate
buffer. `the_system_prefix_stays_byte_stable_across_a_compacting_turn`
still passes.

## Gate fallout

`CompletionRequestRef<'_>` is 3 characters longer than `CompletionRequest`,
which pushes the `complete_ref` signature past 100 columns and costs each
adapter ~3 reflowed lines. `gemini.rs` crossed the 1500-line limit as a
result and is not grandfathered, so its test module moved to
`gemini/tests.rs` — the convention `anthropic.rs` and `zai.rs` already
follow, and `provider_parity`'s source list was repointed at the new path.
Five grandfathered ceilings rise by 1–4 lines in the baseline.
