+++ scripts/file-size-baseline.txt
@@ -20,15 +20,15 @@
 1576 stella-cli/src/candidate_ws.rs
 4578 stella-cli/src/command_deck.rs
 2095 stella-core/src/bus.rs
-2138 stella-core/src/driver.rs
-2820 stella-core/src/driver/tests.rs
+2141 stella-core/src/driver.rs
+2821 stella-core/src/driver/tests.rs
 1849 stella-fleet/src/fleet.rs
 1609 stella-model/src/anthropic/tests.rs
-1780 stella-model/src/bedrock.rs
-1948 stella-model/src/openai.rs
-1684 stella-model/src/zai/tests.rs
+1784 stella-model/src/bedrock.rs
+1952 stella-model/src/openai.rs
+1685 stella-model/src/zai/tests.rs
 2626 stella-pipeline/src/pipeline.rs
-2159 stella-pipeline/src/pipeline/tests.rs
+2162 stella-pipeline/src/pipeline/tests.rs
 2566 stella-protocol/src/event.rs
 2055 stella-store/src/lib.rs
 2204 stella-store/src/tests.rs
