+++ stella-protocol/src/lib.rs
@@ -60,8 +60,8 @@ pub use attachment::{
 };
 pub use cache::CacheCause;
 pub use completion::{
-    CompletionMessage, CompletionRequest, CompletionResult, CompletionUsage, FinishReason,
-    GenerationParams, MessageRole, ReasoningEffort, ServiceTier, Verbosity,
+    CompletionMessage, CompletionRequest, CompletionRequestRef, CompletionResult, CompletionUsage,
+    FinishReason, GenerationParams, MessageRole, ReasoningEffort, ServiceTier, Verbosity,
 };
 pub use context_event::{CompiledContextFrameBuilt, LifecycleEvent, LifecycleEventEnvelope};
 pub use error::ProviderError;
